requirements.txt

streamlit>=1.28.0
pandas>=2.0.0
python-docx>=1.0.0
openai>=1.37.0
anthropic>=0.34.0
Markdown>=3.5.0
PyYAML>=6.0.0
numpy>=1.24.0
pytest>=7.4.0
openpyxl>=3.1.2

python-dotenv>=1.0.1
tiktoken>=0.7.0
